feat: share Connection Between tab socket
